Add the ability to reapply hue and saturation from the original image
Based on https://github.com/AUTOMATIC1111/stable-diffusion-webui-pixelization/pull/23
